+++ packages/orq-rc/src/lib/config.ts
@@ -68,7 +68,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "3.13.0-rc.40",
+  sdkVersion: "3.13.0-rc.41",
   genVersion: "2.716.10",
-  userAgent: "speakeasy-sdk/typescript 3.13.0-rc.40 2.716.10 2.0 @orq-ai/node",
+  userAgent: "speakeasy-sdk/typescript 3.13.0-rc.41 2.716.10 2.0 @orq-ai/node",
 } as const;