+++ packages/orq-rc/src/models/operations/listagents.ts
@@ -1025,7 +1025,7 @@ export const ListAgentsKnowledgeBases$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  id: z.string().default("01K62KZ682MBSC1D49JSVXD139"),
+  id: z.string().default("01K62YGY886SEKTXEF48210N7N"),
   knowledge_id: z.string(),
   configuration: z.union([
     z.lazy(() => ListAgentsConfiguration2$inboundSchema),
@@ -1052,7 +1052,7 @@ export const ListAgentsKnowledgeBases$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   ListAgentsKnowledgeBases
 > = z.object({
-  id: z.string().default("01K62KZ682MBSC1D49JSVXD139"),
+  id: z.string().default("01K62YGY886SEKTXEF48210N7N"),
   knowledgeId: z.string(),
   configuration: z.union([
     z.lazy(() => ListAgentsConfiguration2$outboundSchema),