+++ packages/orq-rc/RELEASES.md
@@ -3879,3 +3879,13 @@ Based on:
 - [typescript v3.13.0-rc.40] packages/orq-rc
 ### Releases
 - [NPM v3.13.0-rc.40] https://www.npmjs.com/package/@orq-ai/node/v/3.13.0-rc.40 - packages/orq-rc
+
+## 2025-09-26 12:05:24
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.630.0 (2.716.10) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.13.0-rc.41] packages/orq-rc
+### Releases
+- [NPM v3.13.0-rc.41] https://www.npmjs.com/package/@orq-ai/node/v/3.13.0-rc.41 - packages/orq-rc